My book ideas
Title:  "Written From My HEART:  I WANT TO MAKE A DIFFERENCE" 
We are born with a  "HEART" Our heart is the begnning of who we will be.  Our heart beats in our mothers womb! When we are first born and come into the world, we cry, so our heart and lungs can begin our life!  A life that is a miracle,  Your heart beats, so you become alive and grow into a child! God gave us our heartbeats so that we become "A child of God"  Our personality becomes shaped by the person we are born to! Beginning life is so different for everyone. It matters who God has given us too! How we are treated as we grown into a child.  Do we have the basics for life? Food, clothing and shelter?  
Children are taught by the person who is providing your care.  That persons personality, beliefs and circumstances have so much to do with how our lives begin! 
Most children are born into loving families, others are shaped by someone who is taking care of you, do they want the demands of a child, can they handle the emotions, responsiblity and have the finacial means to take care of you?Taking care of a child takes alot of heart and love! It is a challenge that some people don't know how to do. How their lives was shaped for their  personality, morals and  understanding about life!  Each person is taught in so many different ways with so many different beliefs and